// src/app/business-logic/model/organization/userDetails.ts
/**
 * organization
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 2.12
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */



export interface UserDetails {
    /**
     * User ID
     */
    id?: string;
    /**
     * User Name
     */
    name?: string;
    /**
     * User avatar (URL or base64-encoded data)
     */
    avatar?: string;
}